changed headers to hpp
