+++ include/entity/squareType.hpp
