+++ include/screen/instructions.hpp
@@ -6,8 +6,8 @@
 #ifndef BATTLESHIP_INSTRUCTIONS_H
 #define BATTLESHIP_INSTRUCTIONS_H
 
-#include "../entity/button.h"
-#include "screenTemplate.h"
+#include "../entity/button.hpp"
+#include "screenTemplate.hpp"
 
 #include <SFML/System.hpp>
