+++ src/helpers/backend.cpp
@@ -2,7 +2,7 @@
  *
  */
 
-#include "../../include/helpers/backend.h"
+#include "../../include/helpers/backend.hpp"
 #include <random>
 
 int randomInt(int start, int end) {
