+++ src/main.cpp
@@ -2,7 +2,7 @@
  * CISC 320 Fall 2021: Atomica group project
  */
 
-#include "../include/screen/screenManager.h"
+#include "../include/screen/screenManager.hpp"
 
 int main() {
     srand(time(NULL));
