+++ src/screen/difficultySelection.cpp
@@ -3,9 +3,9 @@
  * Description: Front-end class that defines the behaviour of the Difficulty selection screen
  */
 
-#include "../../include/screen/difficultySelection.h"
-#include "../../include/helpers/sprites.h"
-#include "../../include/helpers/userInput.h"
+#include "../../include/screen/difficultySelection.hpp"
+#include "../../include/helpers/sprites.hpp"
+#include "../../include/helpers/userInput.hpp"
 
 using screen::DifficultySelection;
