+++ src/screen/fleetPlacement.cpp
@@ -3,11 +3,11 @@
  * Description: Front-end class that defines the behaviour of the Fleet Placement screen
  */
 
-#include "../../include/screen/fleetPlacement.h"
-#include "../../include/helpers/sprites.h"
-#include "../../include/helpers/userInput.h"
-#include "../../include/screen/gameplay.h"
-#include "../../include/helpers/backend.h"
+#include "../../include/screen/fleetPlacement.hpp"
+#include "../../include/helpers/sprites.hpp"
+#include "../../include/helpers/userInput.hpp"
+#include "../../include/screen/gameplay.hpp"
+#include "../../include/helpers/backend.hpp"
 
 using screen::FleetPlacement;
 using std::get;
