+++ src/screen/homepage.cpp
@@ -3,9 +3,9 @@
  * Description: Front-end class that defines the behaviour of the Homepage screen
  */
 
-#include "../../include/screen/homepage.h"
-#include "../../include/helpers/sprites.h"
-#include "../../include/helpers/userInput.h"
+#include "../../include/screen/homepage.hpp"
+#include "../../include/helpers/sprites.hpp"
+#include "../../include/helpers/userInput.hpp"
 
 using screen::Homepage;
