+++ src/screen/intermediary.cpp
@@ -3,9 +3,9 @@
  * Description: Front-end class that defines the behaviour of the Intermediary screen
  */
 
-#include "../../include/screen/intermediary.h"
-#include "../../include/helpers/sprites.h"
-#include "../../include/helpers/userInput.h"
+#include "../../include/screen/intermediary.hpp"
+#include "../../include/helpers/sprites.hpp"
+#include "../../include/helpers/userInput.hpp"
 
 using screen::Intermediary;
