+++ src/screen/screenTemplate.cpp
@@ -2,7 +2,7 @@
  * ScreenTemplate abstract class implementation
  */
 
-#include "../../include/screen/screenTemplate.h"
+#include "../../include/screen/screenTemplate.hpp"
 
 using screen::ScreenTemplate;
